// src/modules/auth/auth.schemas.ts
import {z} from 'zod';

export const registerSchema = z.object({
    name: z.string().min(2).max(100),
    email: z.string().email(),
    password: z.string().min(6).max(100),
    role: z.enum(['customer', 'owner', 'admin']).optional(),
});
export const loginSchema = z.object({
    email: z.string().email(),
    password: z.string().min(6).max(100),
});
export const refreshTokenSchema = z.object({
    refreshToken: z.string().min(10),
});